Revert "Wrapperクラスの関数を各シーン内に移行"
