+++ Application.cpp
@@ -204,9 +204,6 @@ Application& Application::Instance()
 	static Application instance;
 	return instance;
 }
-
-
-
 bool Application::Init()
 {
 	auto result = CoInitializeEx(0, COINIT_MULTITHREADED);
