+++ Pera.h
@@ -4,11 +4,10 @@
 
 
 class Wrapper;
-class TitleScene;
 class Pera
 {
 private:
-	TitleScene& _title;
+
 	struct PeraVertex {
 		DirectX::XMFLOAT3 pos;
 		DirectX::XMFLOAT2 uv;
@@ -29,7 +28,7 @@ private:
 	bool VertexInit();
 
 public:
-	Pera(std::shared_ptr<Wrapper> dx, TitleScene& title);
+	Pera(std::shared_ptr<Wrapper> dx);
 	bool Init();
 	void Draw();
 	~Pera();
